+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.74.11
 
+### Bug Fixes
+
+- Class features added to a character due to gaining levels in a class now get added as linked children to that class, so they get deleted properly when you delete the class (note: items previously added in this way are not affected)
+
 ### Changelog
 
 - Add a notes page to loot sheets.